Current version is 0.16.50: Psi+ was updated. Iris was updated. Libpsi was updated. One plugin was updated. Extra patches were updated.

+++ iris/src/irisnet/noncore/cutestuff/socks.h
@@ -81,8 +81,8 @@ public:
 	void close();
 	void write(const QByteArray &);
 	QByteArray read(int bytes=0);
-	int bytesAvailable() const;
-	int bytesToWrite() const;
+	qint64 bytesAvailable() const;
+	qint64 bytesToWrite() const;
 
 	// remote address
 	QHostAddress peerAddress() const;
@@ -108,7 +108,7 @@ private slots:
 	void sock_connectionClosed();
 	void sock_delayedCloseFinished();
 	void sock_readyRead();
-	void sock_bytesWritten(int);
+	void sock_bytesWritten(qint64);
 	void sock_error(int);
 	void serve();
