+++ iris/src/xmpp/xmpp-core/xmpp_clientstream.h
@@ -171,7 +171,7 @@ namespace XMPP
 		void bs_error(int); // server only
 
 		void ss_readyRead();
-		void ss_bytesWritten(int);
+		void ss_bytesWritten(qint64);
 		void ss_tlsHandshaken();
 		void ss_tlsClosed();
 		void ss_error(int);
